Split utils to different files
